Refactored UI code to create a reusuable component for displaying analysis outputs. Added filtering of single sample analysis outputs. Added iridaworkflowdescription to ProjectSampleAnalysisOutputInfo dto as we need the pipeline name and version for the analysis outputs table
